+++ java/messaging/mqtt/paho-mqtt-client/src/main/java/io/joynr/messaging/mqtt/paho/client/MqttPahoClient.java
@@ -40,7 +40,6 @@ import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
 import edu.umd.cs.findbugs.annotations.SuppressFBWarnings;
-
 import io.joynr.exceptions.JoynrDelayMessageException;
 import io.joynr.exceptions.JoynrIllegalStateException;
 import io.joynr.exceptions.JoynrMessageNotSentException;
@@ -48,7 +47,6 @@ import io.joynr.exceptions.JoynrRuntimeException;
 import io.joynr.messaging.FailureAction;
 import io.joynr.messaging.mqtt.IMqttMessagingSkeleton;
 import io.joynr.messaging.mqtt.JoynrMqttClient;
-import io.joynr.messaging.mqtt.MqttMessagingStub;
 import io.joynr.messaging.mqtt.statusmetrics.MqttStatusReceiver;
 import joynr.system.RoutingTypes.MqttAddress;
 
@@ -186,18 +184,29 @@ public class MqttPahoClient implements JoynrMqttClient, MqttCallback {
                     if (isSecureConnection) {
                         logger.error("Failed to establish TLS connection, error: ", mqttError);
                     }
-                    // fallthrough
+                    // fall through
                 case MqttException.REASON_CODE_BROKER_UNAVAILABLE:
+                    // fall through
                 case MqttException.REASON_CODE_CLIENT_DISCONNECTING:
+                    // fall through
                 case MqttException.REASON_CODE_CLIENT_NOT_CONNECTED:
+                    // fall through
                 case MqttException.REASON_CODE_CLIENT_TIMEOUT:
+                    // fall through
                 case MqttException.REASON_CODE_CONNECT_IN_PROGRESS:
+                    // fall through
                 case MqttException.REASON_CODE_CONNECTION_LOST:
+                    // fall through
                 case MqttException.REASON_CODE_MAX_INFLIGHT:
+                    // fall through
                 case MqttException.REASON_CODE_NO_MESSAGE_IDS_AVAILABLE:
+                    // fall through
                 case MqttException.REASON_CODE_SERVER_CONNECT_ERROR:
+                    // fall through
                 case MqttException.REASON_CODE_SUBSCRIBE_FAILED:
+                    // fall through
                 case MqttException.REASON_CODE_UNEXPECTED_ERROR:
+                    // fall through
                 case MqttException.REASON_CODE_WRITE_TIMEOUT:
                     mqttClientLock.unlockRead(stamp);
                     try {
@@ -220,6 +229,9 @@ public class MqttPahoClient implements JoynrMqttClient, MqttCallback {
                     }
                     stamp = mqttClientLock.readLock();
                     continue;
+                default:
+                    logger.error("Unable to connect for unknown/unlisted reason.");
+                    throw new JoynrIllegalStateException("Unable connect for unknown/unlisted reason.");
                 }
             } catch (JoynrIllegalStateException e) {
                 Throwable cause = e.getCause();
@@ -392,11 +404,6 @@ public class MqttPahoClient implements JoynrMqttClient, MqttCallback {
         } finally {
             mqttClientLock.unlockWrite(stamp);
         }
-    }
-
-    @Override
-    public void publishMessage(String topic, byte[] serializedMessage) {
-        publishMessage(topic, serializedMessage, MqttMessagingStub.DEFAULT_QOS_LEVEL);
     }
 
     @Override