+++ wiki/ReleaseNotes.md
@@ -2,6 +2,18 @@
 All relevant changes are documented in this file. You can find more information about
 the versioning scheme [here](JoynrVersioning.md).
 
+# joynr 1.8.9
+
+## API relevant changes
+None.
+
+## Other changes
+* **[Java]** Fixed and improved reconnect handling and error handling in HivemqMqttClient.
+* **[Java]** Improved logging in HivemqMqttClient.
+
+## Configuration property changes
+None.
+
 # joynr 1.8.8
 
 ## API relevant changes